requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.5.7
distro==1.9.0
frozenlist==1.6.0
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
libcst==1.7.0
magic-filter==1.0.12
marshmallow==3.26.1
marshmallow-enum==1.5.1
mcp==1.6.0
mindee==4.21.0
multidict==6.4.3
mypy_extensions==1.1.0
nodeenv==1.9.1
openai==1.76.0
openai-agents==0.0.13
packaging==25.0
pillow==11.2.1
propcache==0.3.1
psutil==7.0.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pypdfium2==4.30.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
tabulate==0.9.0
TestSlide==2.7.1
tqdm==4.67.1
typeguard==2.13.3
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
